+++ SlashGaming-Diablo-II-Free-MultiBox/resources/resource.rc
@@ -1,8 +1,8 @@
 /**
- * SlashGaming Diablo II MultiBox Patch
- * Copyright (C) 2019  Mir Drualga
+ * SlashGaming Diablo II Free MultiBox
+ * Copyright (C) 2019-2020  Mir Drualga
  *
- * This file is part of SlashGaming Diablo II MultiBox Patch.
+ * This file is part of SlashGaming Diablo II Free MultiBox.
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as published
@@ -69,7 +69,7 @@ VS_VERSION_INFO    VERSIONINFO
       VALUE "Developer",        "Mir Drualga\0"
       VALUE "FileDescription",  "SlashGaming Diablo II MultiBox Patch\0"
       VALUE "InternalName",     "SlashGaming Diablo II MultiBox Patch\0"
-      VALUE "LegalCopyright",   "Copyright (C) 2019  Mir Drualga\0"
+      VALUE "LegalCopyright",   "Copyright (C) 2019-2020  Mir Drualga\0"
       VALUE "LegalTrademarks",  "All rights reserved.\0"
       VALUE "PrivateBuild",     "\0"
       VALUE "ProductName",      "SlashGaming Diablo II Version Detector\0"
