+++ SlashGaming-Diablo-II-Free-MultiBox/src/license.h
@@ -1,8 +1,8 @@
 /**
- * SlashGaming Diablo II MultiBox Patch
- * Copyright (C) 2019  Mir Drualga
+ * SlashGaming Diablo II Free MultiBox
+ * Copyright (C) 2019-2020  Mir Drualga
  *
- * This file is part of SlashGaming Diablo II MultiBox Patch.
+ * This file is part of SlashGaming Diablo II Free MultiBox.
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as published
@@ -35,10 +35,20 @@
  *  work.
  */
 
-#if defined(SGD2MAPI_DLLEXPORT)
-#define DLLEXPORT __declspec(dllexport)
-#elif defined(SGD2MAPI_DLLIMPORT)
-#define DLLEXPORT __declspec(dllimport)
-#else
-#define DLLEXPORT
-#endif
+#ifndef SGD2MBP_LICENSE_H_
+#define SGD2MBP_LICENSE_H_
+
+#include "dllexport_define.inc"
+
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
+DLLEXPORT void PrintLicenseNotice();
+
+#ifdef __cplusplus
+} // extern "C"
+#endif // __cplusplus
+
+#include "dllexport_undefine.inc"
+#endif  // SGD2MBP_LICENSE_H_
